// apps/user-data-service/src/preferences/types/preferences.types.ts
export enum Units {
  METRIC = 'metric',
  IMPERIAL = 'imperial',
}

export enum Theme {
  LIGHT = 'light',
  DARK = 'dark',
}

export enum DateFormat {
  DD_MM_YY_SLASH = 'DD/MM/YYYY',
  MM_DD_YY_SLASH = 'MM/DD/YYYY',
  YYYY_MM_DD_SLASH = 'YYYY/MM/DD',
  DD_MM_YY_DASH = 'DD-MM-YYYY',
  MM_DD_YY_DASH = 'MM-DD-YYYY',
  YYYY_MM_DD_DASH = 'YYYY-MM-DD',
}

export enum TimeFormat {
  TWELVE_HOUR = '12h',
  TWENTY_FOUR_HOUR = '24h',
}